+++ noxfile.py
@@ -62,7 +62,7 @@ def unit_tests(session):
     )
 
 
-@nox.session(python=["3.8"])
+@nox.session(python=["3.9"])
 def isort(session):
     """
     Nox run isort
@@ -81,7 +81,7 @@ def isort(session):
     session.run("isort", "-c", ".")
 
 
-@nox.session(python=["3.8"])
+@nox.session(python=["3.9"])
 def black(session):
     """
     Nox run black
@@ -100,6 +100,7 @@ def black(session):
     session.run("black", "--check", ".")
 
 
+# holding this back to 3.8 due to bug: https://github.com/PyCQA/pylint/issues/3882
 @nox.session(python=["3.8"])
 def pylama(session):
     """
@@ -120,7 +121,7 @@ def pylama(session):
     session.run("pylama", ".")
 
 
-@nox.session(python=["3.8"])
+@nox.session(python=["3.9"])
 def pydocstyle(session):
     """
     Nox run pydocstyle
@@ -139,7 +140,7 @@ def pydocstyle(session):
     session.run("pydocstyle", ".")
 
 
-@nox.session(python=["3.8"])
+@nox.session(python=["3.9"])
 def mypy(session):
     """
     Nox run mypy
@@ -160,7 +161,7 @@ def mypy(session):
     session.run("mypy", "--strict", "scrapli_community/")
 
 
-@nox.session(python=["3.8"])
+@nox.session(python=["3.9"])
 def darglint(session):
     """
     Nox run darglint