+++ scrapli_community/aethra/atosnt/__init__.py
@@ -1,0 +1,4 @@
+"""scrapli_community.aethra.atosnt"""
+from scrapli_community.aethra.atosnt.aethra_atosnt import SCRAPLI_PLATFORM
+
+__all__ = ("SCRAPLI_PLATFORM",)